% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nass_api.R
\name{check_response}
\alias{check_response}
\title{Print human-readable messages for http errors.}
\usage{
check_response(status_code)
}
\arguments{
\item{status_code}{The http response code.}
}
\value{
Nothing.
}
\description{
Print human-readable messages for http errors.
}
