% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nass_api.R
\name{get_county_data}
\alias{get_county_data}
\title{Get the data for the specified query for county-level data.}
\usage{
get_county_data(key, year, data_item, fips = "all", domain = "TOTAL")
}
\arguments{
\item{key}{Your NASS API key.}

\item{year}{Must be a census year (e.g. 2012, 2017).}

\item{data_item}{The long description of the desired series. Use
search_data_items function to find one.}

\item{fips}{Must be 'all', a 2-digit state fips, or a 5-digit county fips.}

\item{domain}{A modifier on data_item, some characterstic (e.g. size
categories of operations), use 'all' to get all.}
}
\value{
A tibble df of the requested data, if any exists. Otherwise returns
NA.
}
\description{
Get the data for the specified query for county-level data.
}
\examples{
key <- Sys.getenv('NASS_KEY')
get_county_data(key=key, year=2017,
                data_item='CORN, GRAIN - ACRES HARVESTED', fips='all')
get_county_data(key=key, year=2017,
                data_item='CORN, GRAIN - ACRES HARVESTED', fips='08')
get_county_data(key=key, year=2017,
                data_item='CORN, GRAIN - ACRES HARVESTED', fips='08069',
                domain='all')
}
