% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nass_api.R
\name{get_param_values}
\alias{get_param_values}
\title{Get all values a parameter can take.}
\usage{
get_param_values(param)
}
\arguments{
\item{param}{The parameter name.}
}
\value{
A list of all values that the parameter can take.
}
\description{
Get all values of a parameters that can be passed in a GET request.
See \url{https://quickstats.nass.usda.gov/api} for a table of parameter names.
}
\examples{
get_param_values('source_desc')
get_param_values('domain_desc')
}
