% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nass_api.R
\name{get_state_item_count}
\alias{get_state_item_count}
\title{Get the count of values that exist for the specified query for state-level
data.}
\usage{
get_state_item_count(
  key,
  year,
  data_item,
  fips = "all",
  domain = "TOTAL",
  source = "CENSUS"
)
}
\arguments{
\item{key}{Your NASS API key.}

\item{year}{Must be a census year (e.g. 2012, 2017).}

\item{data_item}{The long description of the desired series. Use
search_data_items function to find one.}

\item{fips}{Must be 'all' or a 2-digit state fips.}

\item{domain}{A modifier on data_item, some characterstic (e.g. size
categories of operations), use 'all' to get all.}

\item{source}{Must be 'CENSUS' or 'SURVEY'.}
}
\value{
The count of values.
}
\description{
This is used as a utility function by other functions, but can also be used
to explore expected results before pulling real data.
}
\examples{
\donttest{
key <- Sys.getenv('NASS_KEY')
get_state_item_count(key=key, year=2017,
                     data_item='CORN, GRAIN - ACRES HARVESTED', fips='all')
get_state_item_count(key=key, year=2017,
                     data_item='CORN, GRAIN - ACRES HARVESTED', fips='08')
}
}
